Add record workflow created, with the corresponding propagation (subcategory amount and account balance)

+++ api/models/balance.go
@@ -1,7 +1,7 @@
 package models
 
 type Balance struct {
-	ID              int64   `gorm:"primary_key;auto_increment" json:"id"`
+	ID              int     `gorm:"primary_key;auto_increment" json:"id"`
 	AvailableAmount float64 `json:"available_amount"`
 	SavingsAmount   float64 `json:"savings_amount"`
 	Debt            float64 `json:"debt"`
