+++ api/models/category.go
@@ -1,7 +1,7 @@
 package models
 
 type Category struct {
-	ID        int64   `gorm:"primary_key;auto_increment" json:"id"`
+	ID        int     `gorm:"primary_key;auto_increment" json:"id"`
 	Name      string  `json:"name"`
 	Income    bool    `json:"income"`
 	AccountID int     `json:"-"`
